requirements.txt
Django==1.6.2
Pillow==2.4.0
South==0.8.4
WebOb==1.3.1
WebTest==2.0.14
argparse==1.2.1
beautifulsoup4==4.3.2
behave==1.2.4
dj-database-url==0.3.0
dj-static==0.0.5
django-behave==0.1.1
django-downtime==0.3.1
django-embed-video==0.8
django-extensions==1.3.3
django-toolbelt==0.0.1
django-webtest==1.7.6
enum34==0.9.23
gunicorn==18.0
ipdb==0.8
ipython==1.2.1
parse==1.6.3
parse-type==0.3.4
psycopg2==2.5.2
pystache==0.5.3
requests==2.2.1
selenium==2.41.0
six==1.6.1
static==1.0.2
waitress==0.8.8
wsgiref==0.1.2
https://github.com/silverfix/django-nested-inlines/archive/master.zip
django-autoslug==1.7.2

